// src/generated/models/Eamon_nbaplayerstatsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Eamon_nbaplayerstatseseamon_is_drafted = { 0: "No", 1: "Yes" } as const;
export type Eamon_nbaplayerstatseseamon_is_drafted = keyof typeof Eamon_nbaplayerstatseseamon_is_drafted;
export const Eamon_nbaplayerstatsesstatecode = { 0: "Active", 1: "Inactive" } as const;
export type Eamon_nbaplayerstatsesstatecode = keyof typeof Eamon_nbaplayerstatsesstatecode;
export const Eamon_nbaplayerstatsesstatuscode = { 1: "Active", 2: "Inactive" } as const;
export type Eamon_nbaplayerstatsesstatuscode = keyof typeof Eamon_nbaplayerstatsesstatuscode;

export interface Eamon_nbaplayerstatsesBase {
  eamon_birth_country?: string;
  eamon_espn_id?: string;
  eamon_fantasy_avg_total?: string;
  eamon_fantasy_avg_yr1?: string;
  eamon_fantasy_avg_yr2?: string;
  eamon_fantasy_avg_yr3?: string;
  eamon_full_name: string;
  eamon_games_pct_total?: string;
  eamon_games_pct_yr1?: string;
  eamon_games_pct_yr2?: string;
  eamon_games_pct_yr3?: string;
  eamon_is_drafted?: Eamon_nbaplayerstatseseamon_is_drafted;
  eamon_last_updated?: string;
  eamon_name?: string;
  eamon_nbaplayerstatsid: string;
  eamon_player_id: string;
  eamon_position?: string;
  eamon_team?: string;
  eamon_weekly_high_avg_total?: string;
  eamon_weekly_high_avg_yr1?: string;
  eamon_weekly_high_avg_yr2?: string;
  eamon_weekly_high_avg_yr3?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid?: string;
  owneridtype?: string;
  statecode: Eamon_nbaplayerstatsesstatecode;
  statuscode?: Eamon_nbaplayerstatsesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Eamon_nbaplayerstatses extends Eamon_nbaplayerstatsesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  eamon_is_draftedname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}